// src/types/types.ts
interface DataType {
    ou: string
    year: number
    sport: String
    home: Boolean
    division: Boolean
    g1FavTwoPointFive: string
    g1FavOnePointFiveThreePointFive: string
    g1DogTwoPointFive: string
    g1DogOnePointFiveThreePointFive: string
    favTwoPointFive: string
    favOnePointFiveThreePointFive: string
    dogTwoPointFive: string
    dogOnePointFiveThreePointFive: string
    homeMLTwoPointFive: string
    homeMLThreePointFive: string
}

export type { DataType }